+++ ft_memchr.c
@@ -9,7 +9,7 @@ void *ft_memchr(const void *s, int c, size_t n)
 	i = -1;
 	s_ucp = (unsigned char *)s;
 	c_uc = (unsigned char)c;
-	while(++i < n)
+	while(++i < (int)n)
 		if(s_ucp[i] == c_uc)
 			return (void *)s_ucp + i;
