+++ ft_strncmp.c
@@ -5,7 +5,7 @@ int ft_strncmp(const char *s1, const char *s2, size_t n)
 	int i;
 
 	i = -1;
-	while(++i < n)
+	while(++i <	(int)n)
 		if (s1[i] != s2[i])
 			return (s1[i] - '0') - (s2[i] - '0');
 	return 0;
